requirements.txt
pathlib2
python-dotenv
requests
python-dotenv
atlassian-python-api